backend/requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.6
cryptography
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.35
alembic==1.14.0
httpx==0.27.2
pyjwt==2.9.0
bcrypt==4.2.1
python-dotenv==1.0.1
pydantic-settings==2.6.0
